+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Default environment to "debug" if running with debugger attached (#978)
 
+### Fixes
+
+- Fixed an issue where an attempt to deserialize `Device` with a non-system time zone failed (#993)
+
 ## 3.3.4
 
 ### Features